requirements.txt
python-docx==0.8.11
openai>=1.3.7
sentence-transformers
pandas
openpyxl 